requirements.txt
dj-database-url==0.2.2
django-extensions==3.2.1
Django==5.0.4
djangorestframework==3.15.1
psycopg==3.1.17


# Development
black==23.7.0
pylint==3.1.0
Werkzeug==3.0.1
